+++ tools/audit_tools.py
@@ -3,6 +3,7 @@
 from typing import Any
 
 from utils.api_call import http_call_response
+from utils.common import build_list_params
 from utils.decorators import mcp_tool_handler
 from utils.http_client import http_client
 from utils.tool_annotations import READ_ONLY
@@ -37,11 +38,7 @@ async def list_activity_logs(
     """
     token = kwargs.get('token')
 
-    params = {}
-    if page is not None:
-        params['page'] = page
-    if page_size is not None:
-        params['page_size'] = page_size
+    params = build_list_params(page=page, page_size=page_size)
 
     return await http_call_response(
         http_client.get,
@@ -117,13 +114,7 @@ async def list_server_logs(
     """
     token = kwargs.get('token')
 
-    params: dict[str, Any] = {}
-    if server_id:
-        params['server'] = server_id
-    if page is not None:
-        params['page'] = page
-    if page_size is not None:
-        params['page_size'] = page_size
+    params = build_list_params(page=page, page_size=page_size, server=server_id)
 
     return await http_call_response(
         http_client.get,
@@ -164,13 +155,7 @@ async def list_webftp_logs(
     """
     token = kwargs.get('token')
 
-    params: dict[str, Any] = {}
-    if server_id:
-        params['server'] = server_id
-    if page is not None:
-        params['page'] = page
-    if page_size is not None:
-        params['page_size'] = page_size
+    params = build_list_params(page=page, page_size=page_size, server=server_id)
 
     return await http_call_response(
         http_client.get,
@@ -220,17 +205,13 @@ async def list_session_analyses(
     """
     token = kwargs.get('token')
 
-    params: dict[str, Any] = {}
-    if server_id:
-        params['server'] = server_id
-    if status:
-        params['status'] = status
-    if risk_score:
-        params['risk_score'] = risk_score
-    if page is not None:
-        params['page'] = page
-    if page_size is not None:
-        params['page_size'] = page_size
+    params = build_list_params(
+        page=page,
+        page_size=page_size,
+        server=server_id,
+        status=status,
+        risk_score=risk_score,
+    )
 
     return await http_call_response(
         http_client.get,