+++ tools/cert_tools.py
@@ -3,7 +3,7 @@
 from typing import Any
 
 from utils.api_call import http_call_response
-from utils.common import error_response
+from utils.common import build_list_params, error_response
 from utils.decorators import mcp_tool_handler
 from utils.http_client import http_client
 from utils.tool_annotations import ADDITIVE, DESTRUCTIVE, IDEMPOTENT_WRITE, READ_ONLY
@@ -38,11 +38,7 @@ async def list_certificate_authorities(
     """
     token = kwargs.get('token')
 
-    params: dict[str, Any] = {}
-    if page is not None:
-        params['page'] = page
-    if page_size is not None:
-        params['page_size'] = page_size
+    params = build_list_params(page=page, page_size=page_size)
 
     return await http_call_response(
         http_client.get,
@@ -280,11 +276,7 @@ async def list_sign_requests(
     """
     token = kwargs.get('token')
 
-    params: dict[str, Any] = {}
-    if page is not None:
-        params['page'] = page
-    if page_size is not None:
-        params['page_size'] = page_size
+    params = build_list_params(page=page, page_size=page_size)
 
     return await http_call_response(
         http_client.get,
@@ -557,13 +549,11 @@ async def list_certificates(
     """
     token = kwargs.get('token')
 
-    params: dict[str, Any] = {}
-    if authority_id is not None:
-        params['csr__authority__id'] = authority_id
-    if page is not None:
-        params['page'] = page
-    if page_size is not None:
-        params['page_size'] = page_size
+    params = build_list_params(
+        page=page,
+        page_size=page_size,
+        csr__authority__id=authority_id,
+    )
 
     return await http_call_response(
         http_client.get,
@@ -691,11 +681,7 @@ async def list_revoke_requests(
     """
     token = kwargs.get('token')
 
-    params: dict[str, Any] = {}
-    if page is not None:
-        params['page'] = page
-    if page_size is not None:
-        params['page_size'] = page_size
+    params = build_list_params(page=page, page_size=page_size)
 
     return await http_call_response(
         http_client.get,